// frontend/src/api/models/Wahl.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bundestagswahlsystem API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Wahl
 */
export interface Wahl {
    /**
     * 
     * @type {number}
     * @memberof Wahl
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Wahl
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof Wahl
     */
    date: Date;
}

/**
 * Check if a given object implements the Wahl interface.
 */
export function instanceOfWahl(value: object): value is Wahl {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('date' in value) || value['date'] === undefined) return false;
    return true;
}

export function WahlFromJSON(json: any): Wahl {
    return WahlFromJSONTyped(json, false);
}

export function WahlFromJSONTyped(json: any, ignoreDiscriminator: boolean): Wahl {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'date': (new Date(json['date'])),
    };
}

export function WahlToJSON(json: any): Wahl {
    return WahlToJSONTyped(json, false);
}

export function WahlToJSONTyped(value?: Wahl | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'date': ((value['date']).toISOString().substring(0,10)),
    };
}