// frontend/src/api/models/WahlzettelParteiWrapper.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bundestagswahlsystem API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Abgeordneter } from './Abgeordneter';
import {
    AbgeordneterFromJSON,
    AbgeordneterFromJSONTyped,
    AbgeordneterToJSON,
    AbgeordneterToJSONTyped,
} from './Abgeordneter';
import type { Partei } from './Partei';
import {
    ParteiFromJSON,
    ParteiFromJSONTyped,
    ParteiToJSON,
    ParteiToJSONTyped,
} from './Partei';

/**
 * 
 * @export
 * @interface WahlzettelParteiWrapper
 */
export interface WahlzettelParteiWrapper {
    /**
     * 
     * @type {Partei}
     * @memberof WahlzettelParteiWrapper
     */
    partei: Partei;
    /**
     * 
     * @type {Array<Abgeordneter>}
     * @memberof WahlzettelParteiWrapper
     */
    topfive: Array<Abgeordneter>;
}

/**
 * Check if a given object implements the WahlzettelParteiWrapper interface.
 */
export function instanceOfWahlzettelParteiWrapper(value: object): value is WahlzettelParteiWrapper {
    if (!('partei' in value) || value['partei'] === undefined) return false;
    if (!('topfive' in value) || value['topfive'] === undefined) return false;
    return true;
}

export function WahlzettelParteiWrapperFromJSON(json: any): WahlzettelParteiWrapper {
    return WahlzettelParteiWrapperFromJSONTyped(json, false);
}

export function WahlzettelParteiWrapperFromJSONTyped(json: any, ignoreDiscriminator: boolean): WahlzettelParteiWrapper {
    if (json == null) {
        return json;
    }
    return {
        
        'partei': ParteiFromJSON(json['partei']),
        'topfive': ((json['topfive'] as Array<any>).map(AbgeordneterFromJSON)),
    };
}

export function WahlzettelParteiWrapperToJSON(json: any): WahlzettelParteiWrapper {
    return WahlzettelParteiWrapperToJSONTyped(json, false);
}

export function WahlzettelParteiWrapperToJSONTyped(value?: WahlzettelParteiWrapper | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'partei': ParteiToJSON(value['partei']),
        'topfive': ((value['topfive'] as Array<any>).map(AbgeordneterToJSON)),
    };
}